+++ src/lib.rs
@@ -1,1 +1,6 @@
 #![doc = include_str!("../README.md")]
+
+mod directory;
+pub use directory::Directory;
+
+pub mod util;